Update SOS DP.cpp

+++ Algorithms-and-Common-Problems/Dynamic Programming/SOS DP.cpp
@@ -3,22 +3,21 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-const int N = (1 << 22)+5;
-int dp[N];
+const ll N = (1 << 22)+5;
+ll dp[N];
 
-signed main()
-{
-    int n; cin >> n;
+signed main() {
+    ll n; cin >> n;
 
-    for (int i = 0; i < N; i++) dp[i] = -1;
-    int arr[n];
+    for (ll i = 0; i < N; i++) dp[i] = -1;
+    ll arr[n];
     for (auto &u : arr) {
         cin >> u;
         dp[u] = u;
     }
 
-    for (int mask = 0; mask < N; mask++) {
-        for (int i = 0; i < 22; i++) {
+    for (ll mask = 0; mask < N; mask++) {
+        for (ll i = 0; i < 22; i++) {
             if (mask & (1 << i)) 
                 dp[mask] = max(dp[mask], dp[(mask ^ (1<<i))]);
         }
@@ -52,8 +51,7 @@ ll func(ll mask, ll i) {
     return dp[mask][i] = ans;
 }
 
-signed main()
-{
+signed main() {
     ll n; cin >> n;
     ll arr[n];
     for (auto &u : arr) {
